Move plot_images tests into plot_signal2d; update and add plot test images
